CHANGES! WE ARE HERE!

+++ project/Move.java
@@ -14,7 +14,7 @@ public abstract class Move {
     // Metadata for proper function
     protected final boolean isActiveWhite;
     protected final byte halfMoveClock;
-    protected final byte moveCounter;
+    protected final int moveCounter;
     protected final byte[] enPassant;
     protected final boolean canWhiteCastleQueen;
     protected final boolean canWhiteCastleKing;
@@ -29,7 +29,7 @@ public abstract class Move {
      * @param toPositionX   The X position where the piece should go.
      * @param toPositionY   The Y position where the piece should go.
      */
-    public Move(byte fromPositionX, byte fromPositionY, byte toPositionX, byte toPositionY, byte fromPiece, byte toPiece, boolean isActiveWhite, byte halfMoveClock, byte moveCounter, byte[] enPassant, boolean canWhiteCastleQueen, boolean canWhiteCastleKing, boolean canBlackCastleQueen, boolean canBlackCastleKing) {
+    public Move(byte fromPositionX, byte fromPositionY, byte toPositionX, byte toPositionY, byte fromPiece, byte toPiece, boolean isActiveWhite, byte halfMoveClock, int moveCounter, byte[] enPassant, boolean canWhiteCastleQueen, boolean canWhiteCastleKing, boolean canBlackCastleQueen, boolean canBlackCastleKing) {
         this.fromPositionX = fromPositionX;
         this.fromPositionY = fromPositionY;
         this.toPositionX = toPositionX;
